Refactor telemetry integration across various modules
- Updated FieldTransformTool to use MetricsTagList for telemetry tagging.
- Refactored ClassificationTreeCapture to utilize MetricsTagList for metrics.
- Changed NodeEnsurer to employ MetricsTagList for telemetry metrics.
- Modified NodeTranslationTool to use MetricsTagList for operation tagging.
- Enhanced TfsClassificationTreeCapture to support metrics recording with MetricsTagList.
- Adjusted DependencyDiscoveryModuleResumeTests to include IDependencyOrchestrator in constructor.
- Updated IdentitiesModuleTests to include IIdentitiesOrchestrator in constructor.
- Refactored NodesModuleTests to utilize INodesOrchestrator in constructor.
- Changed TeamsModuleTests to replace slugGenerator with orchestrator in various instances.
- Updated PackageConfigStoreTests to verify metrics with MetricsTagList.
- Refactored DiscoveryMetricsTests to utilize MetricsTagList for tagging.
- Adjusted MigrationMetricsTests to use MetricsTagList for test tags.
