+++ src/DevOpsMigrationPlatform.Abstractions.Agent/Modules/INodesOrchestrator.cs
@@ -1,0 +1,41 @@
+using System.Threading;
+using System.Threading.Tasks;
+using DevOpsMigrationPlatform.Abstractions.Agent.Checkpointing;
+using DevOpsMigrationPlatform.Abstractions.Agent.Context;
+using DevOpsMigrationPlatform.Abstractions.Agent.Export;
+using DevOpsMigrationPlatform.Abstractions.Agent.Import;
+using DevOpsMigrationPlatform.Abstractions.Agent.Storage;
+using DevOpsMigrationPlatform.Abstractions.Agent.Tools;
+using DevOpsMigrationPlatform.Abstractions.Agent.Validation;
+using DevOpsMigrationPlatform.Abstractions.Validation;
+
+namespace DevOpsMigrationPlatform.Abstractions.Agent.Modules;
+
+/// <summary>
+/// Orchestrates classification-tree (node) export, import, and validation operations.
+/// </summary>
+public interface INodesOrchestrator
+{
+    Task ExportAsync(
+        IClassificationTreeCapture capture,
+        ExportContext context,
+        ISourceEndpointInfo sourceEndpointInfo,
+        ICheckpointingServiceFactory? checkpointingFactory,
+        CancellationToken ct);
+
+#if !NET481
+    Task ImportAsync(
+        INodeEnsurer nodeEnsurer,
+        ImportContext context,
+        ISourceEndpointInfo sourceEndpointInfo,
+        ITargetEndpointInfo targetEndpointInfo,
+        ICheckpointingServiceFactory? checkpointingFactory,
+        bool replicateSourceTree,
+        CancellationToken ct);
+#endif
+
+    Task ValidateAsync(
+        IArtefactStore artefactStore,
+        ValidationContext context,
+        CancellationToken ct);
+}
