+++ src/DevOpsMigrationPlatform.Abstractions.Agent/Tools/IClassificationTreeCapture.cs
@@ -1,9 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using DevOpsMigrationPlatform.Abstractions.Agent.Storage;
-#if !NET481
 using DevOpsMigrationPlatform.Abstractions.Agent.Telemetry;
-#endif
 using DevOpsMigrationPlatform.Abstractions.Streaming;
 
 namespace DevOpsMigrationPlatform.Abstractions.Agent.Tools;
@@ -19,12 +17,10 @@ public interface IClassificationTreeCapture
     /// <returns>Total number of nodes captured (area + iteration).</returns>
     Task<int> CaptureAsync(
         IArtefactStore artefactStore,
-        CancellationToken ct
-#if !NET481
-        , IMigrationMetrics? metrics = null
-        , string? jobId = null
-        , IProgressSink? sink = null
-        , string moduleName = "Nodes"
-#endif
+        CancellationToken ct,
+        IMigrationMetrics? metrics = null,
+        string? jobId = null,
+        IProgressSink? sink = null,
+        string moduleName = "Nodes"
     );
 }
