+++ src/DevOpsMigrationPlatform.Infrastructure.Agent/Modules/TeamsModule.cs
@@ -10,11 +10,9 @@ using DevOpsMigrationPlatform.Abstractions.Agent.Export;
 using DevOpsMigrationPlatform.Abstractions.Agent.Import;
 using DevOpsMigrationPlatform.Abstractions.Agent.Modules;
 using DevOpsMigrationPlatform.Abstractions.Agent.Teams;
-using DevOpsMigrationPlatform.Abstractions.Agent.Telemetry;
 using DevOpsMigrationPlatform.Abstractions.Agent.Tools;
 using DevOpsMigrationPlatform.Abstractions.Agent.Validation;
 using DevOpsMigrationPlatform.Abstractions.Validation;
-using DevOpsMigrationPlatform.Infrastructure.Agent.Teams;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
@@ -37,15 +35,12 @@ public sealed class TeamsModule : IModule
 {
     private readonly ITeamSource? _teamSource;
     private readonly ITeamTarget? _teamTarget;
-    private readonly TeamExportOrchestrator? _exportOrchestrator;
-    private readonly TeamImportOrchestrator? _importOrchestrator;
-    private readonly TeamSlugGenerator _slugGenerator;
     private readonly ICheckpointingServiceFactory? _checkpointingFactory;
     private readonly ILogger<TeamsModule> _logger;
     private readonly TeamsModuleOptions _options;
     private readonly ISourceEndpointInfo _sourceEndpointInfo;
     private readonly ITargetEndpointInfo _targetEndpointInfo;
-    private readonly TeamsOrchestrator _orchestrator;
+    private readonly ITeamsOrchestrator _orchestrator;
 
     public string Name => "Teams";
     public IReadOnlyList<ModuleDependency> DependsOn => new[]
@@ -61,25 +56,19 @@ public sealed class TeamsModule : IModule
         IOptions<TeamsModuleOptions> options,
         ISourceEndpointInfo sourceEndpointInfo,
         ITargetEndpointInfo targetEndpointInfo,
-        TeamSlugGenerator slugGenerator,
+        ITeamsOrchestrator orchestrator,
         ITeamSource? teamSource = null,
         ITeamTarget? teamTarget = null,
-        TeamExportOrchestrator? exportOrchestrator = null,
-        TeamImportOrchestrator? importOrchestrator = null,
-        ICheckpointingServiceFactory? checkpointingFactory = null,
-        IMigrationMetrics? migrationMetrics = null)
+        ICheckpointingServiceFactory? checkpointingFactory = null)
     {
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
         _sourceEndpointInfo = sourceEndpointInfo ?? throw new ArgumentNullException(nameof(sourceEndpointInfo));
         _targetEndpointInfo = targetEndpointInfo ?? throw new ArgumentNullException(nameof(targetEndpointInfo));
-        _slugGenerator = slugGenerator ?? throw new ArgumentNullException(nameof(slugGenerator));
+        _orchestrator = orchestrator ?? throw new ArgumentNullException(nameof(orchestrator));
         _teamSource = teamSource;
         _teamTarget = teamTarget;
-        _exportOrchestrator = exportOrchestrator;
-        _importOrchestrator = importOrchestrator;
         _checkpointingFactory = checkpointingFactory;
-        _orchestrator = new TeamsOrchestrator(logger, migrationMetrics);
     }
 
     /// <inheritdoc/>
@@ -97,15 +86,8 @@ public sealed class TeamsModule : IModule
             return;
         }
 
-        if (_exportOrchestrator is null)
-        {
-            _logger.LogWarning("[Teams] No TeamExportOrchestrator registered — team export skipped.");
-            return;
-        }
-
         await _orchestrator.ExportAsync(
-            _teamSource, _exportOrchestrator, _slugGenerator,
-            context, _sourceEndpointInfo, _checkpointingFactory,
+            _teamSource, context, _sourceEndpointInfo, _checkpointingFactory,
             _options, ct).ConfigureAwait(false);
     }
 
@@ -124,14 +106,8 @@ public sealed class TeamsModule : IModule
             return;
         }
 
-        if (_importOrchestrator is null)
-        {
-            _logger.LogWarning("[Teams] No TeamImportOrchestrator registered — team import skipped.");
-            return;
-        }
-
         await _orchestrator.ImportAsync(
-            _importOrchestrator, context, _sourceEndpointInfo, _targetEndpointInfo,
+            context, _sourceEndpointInfo, _targetEndpointInfo,
             _checkpointingFactory, _options, ct).ConfigureAwait(false);
     }
