+++ src/DevOpsMigrationPlatform.Infrastructure.Agent/Storage/PackageConfigStore.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using DevOpsMigrationPlatform.Abstractions;
+using DevOpsMigrationPlatform.Abstractions.Telemetry;
 using DevOpsMigrationPlatform.Abstractions.Agent.Lease;
 using DevOpsMigrationPlatform.Abstractions.Agent.Storage;
 using DevOpsMigrationPlatform.Abstractions.Agent.Telemetry;
@@ -79,22 +80,19 @@ internal sealed class PackageConfigStore : IPackageConfigStore
                 .ConfigureAwait(false);
 
             sw.Stop();
-            var emptyTags = default(TagList);
-            _metrics?.RecordConfigWriteCompleted(in emptyTags);
+            _metrics?.RecordConfigWriteCompleted(MetricsTagList.Empty);
             activity?.SetTag("outcome", "success");
             _logger.LogInformation("Config copied to package in {DurationMs}ms", sw.ElapsedMilliseconds);
         }
         catch (InvalidOperationException)
         {
-            var emptyTags = default(TagList);
-            _metrics?.RecordConfigWriteError(in emptyTags);
+            _metrics?.RecordConfigWriteError(MetricsTagList.Empty);
             activity?.SetTag("outcome", "exists_error");
             throw;
         }
         catch (Exception)
         {
-            var emptyTags = default(TagList);
-            _metrics?.RecordConfigWriteError(in emptyTags);
+            _metrics?.RecordConfigWriteError(MetricsTagList.Empty);
             activity?.SetTag("outcome", "error");
             _logger.LogError("Failed to copy config to package");
             throw;
@@ -132,9 +130,9 @@ internal sealed class PackageConfigStore : IPackageConfigStore
 
         if (!exists)
         {
-            var emptyTags = default(TagList);
-            _metrics?.RecordConfigReadError(in emptyTags);
-            _metrics?.RecordConfigReadFallback(in emptyTags);
+            var empty = MetricsTagList.Empty;
+            _metrics?.RecordConfigReadError(empty);
+            _metrics?.RecordConfigReadFallback(empty);
             activity?.SetTag("outcome", "not_found");
             _logger.LogWarning(
                 "migration-config.json not found in package after retries. " +
@@ -149,8 +147,7 @@ internal sealed class PackageConfigStore : IPackageConfigStore
 
             if (string.IsNullOrWhiteSpace(json))
             {
-                var emptyTags = default(TagList);
-                _metrics?.RecordConfigReadError(in emptyTags);
+                _metrics?.RecordConfigReadError(MetricsTagList.Empty);
                 activity?.SetTag("outcome", "empty");
                 _logger.LogError("migration-config.json is empty in package");
                 throw new InvalidOperationException(
@@ -165,8 +162,7 @@ internal sealed class PackageConfigStore : IPackageConfigStore
                 .Build();
 
             sw.Stop();
-            var emptyTags2 = default(TagList);
-            _metrics?.RecordConfigReadCompleted(in emptyTags2);
+            _metrics?.RecordConfigReadCompleted(MetricsTagList.Empty);
             activity?.SetTag("outcome", "success");
             _logger.LogInformation("Config read from package in {DurationMs}ms", sw.ElapsedMilliseconds);
 
@@ -178,8 +174,7 @@ internal sealed class PackageConfigStore : IPackageConfigStore
         }
         catch (Exception)
         {
-            var emptyTags = default(TagList);
-            _metrics?.RecordConfigReadError(in emptyTags);
+            _metrics?.RecordConfigReadError(MetricsTagList.Empty);
             activity?.SetTag("outcome", "parse_error");
             _logger.LogError("Failed to parse migration-config.json");
             throw;
